# Performance: add e2e tests for load and typing time