--- a/packages/e2e-tests/jest.performance.config.js
+++ b/packages/e2e-tests/jest.performance.config.js
@@ -1,15 +1,21 @@
 module.exports = {
 	...require( '@wordpress/scripts/config/jest-e2e.config' ),
+	testMatch: [
+		'**/performance.test.js',
+	],
 	setupFiles: [
 		'<rootDir>/config/gutenberg-phase.js',
 	],
 	setupFilesAfterEnv: [
 		'<rootDir>/config/setup-test-framework.js',
-		'@wordpress/jest-console',
-		'@wordpress/jest-puppeteer-axe',
 		'expect-puppeteer',
 	],
-	testPathIgnorePatterns: [
-		'e2e-tests/specs/performance.test.js',
+	transformIgnorePatterns: [
+		'node_modules',
+		'scripts/config/puppeteer.config.js',
+	],
+	reporters: [
+		'default',
+		'<rootDir>/config/performance-reporter.js',
 	],
 };