--- a/packages/e2e-tests/jest.performance.config.js
+++ b/packages/e2e-tests/jest.performance.config.js
@@ -8,6 +8,8 @@ module.exports = {
 	],
 	setupFilesAfterEnv: [
 		'<rootDir>/config/setup-test-framework.js',
+		'@wordpress/jest-console',
+		'@wordpress/jest-puppeteer-axe',
 		'expect-puppeteer',
 	],
 	transformIgnorePatterns: [