--- a/packages/e2e-tests/specs/performance.test.js
+++ b/packages/e2e-tests/specs/performance.test.js
@@ -19,7 +19,7 @@ function readFile( filePath ) {
 
 describe( 'Performance', () => {
 	it( '1000 paragraphs', async () => {
-		const html = readFile( join( __dirname, '../assets/neuralink.html' ) );
+		const html = readFile( join( __dirname, '../assets/large-post.html' ) );
 
 		await createNewPost();
 		await page.evaluate( ( _html ) => {