… packages/e2e-tests/assets/large-post.html
… packages/e2e-tests/assets/neuralink.html
… packages/e2e-tests/specs/.gitignore