## Description

Adds some simple tests to report the average load and typing times.

Example:

```
Average time to load: 4889.1ms
Average time to DOM content load: 4034.9ms
Average time to type character: 67.165ms
Slowest time to type character: 140ms
Fastest time to type character: 50ms
```

Why? It is more accurate than trying to record key presses at random places in a big post. It's also possible to test with different content. Note that speed can still vary depending on a number of environmental factors, but at least some things are now constant.

The tests are done with a 1000 paragraph post, where each paragraph contains 200 characters. Maybe additional tests should be made with more diverse content, but I think this is a good start.

For load times, 10 samples are taken. For typing speed 200 samples are taken (typing one full paragraph).

The typing time includes everything Puppeteer is executing during `page.keyboard.type`, so, in reality, it is probably less. We can try to look at `keyup` and `keydown` events instead.

## How has this been tested?
<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, tests ran to see how -->
<!-- your change affects other areas of the code, etc. -->

## Screenshots <!-- if applicable -->

## Types of changes
<!-- What types of changes does your code introduce?  -->
<!-- Bug fix (non-breaking change which fixes an issue) -->
<!-- New feature (non-breaking change which adds functionality) -->
<!-- Breaking change (fix or feature that would cause existing functionality to not work as expected) -->

## Checklist:
- [ ] My code is tested.
- [ ] My code follows the WordPress code style. <!-- Check code: `npm run lint`, Guidelines: https://make.wordpress.org/core/handbook/best-practices/coding-standards/javascript/ -->
- [ ] My code follows the accessibility standards. <!-- Guidelines: https://make.wordpress.org/core/handbook/best-practices/coding-standards/accessibility-coding-standards/ -->
- [ ] My code has proper inline documentation. <!-- Guidelines: https://make.wordpress.org/core/handbook/best-practices/inline-documentation-standards/javascript/ -->
- [ ] I've included developer documentation if appropriate. <!-- Handbook: https://wordpress.org/gutenberg/handbook/designers-developers/ -->