@youknowriad 2019-03-19
Is there a way we could automate this across versions. Like have a shell script that checkout old tags, runs this test and produce a table at the end?