@youknowriad 2019-03-19
Can we test in a post with images, quotes... like this one for instance https://waitbutwhy.com/2017/04/neuralink.html