@ellatrix 2019-03-19
@youknowriad I tried checking out other tags and running the tests, the only problem is that the tests don't exist in the working directory then. Not sure how to solve it. I can try copying the file, then writing it for each tag.