@youknowriad 2019-03-19
@ellatrix I was thinking we use the exact same test in several checkouts:

 - Keep the current set of tests (master)
 - Checkout the tags in a different temporary folder
 - Prepare the environment (reinstall packages, rebuild...)
 - Run the current set of tests on this environment.