@ellatrix 2019-03-19
I'm slowly giving up on this :(

Always keep running into vague errors

> Validation Error:
>
>  Options: setupTestFrameworkScriptFile and setupFilesAfterEnv cannot be used together.
>  Please change your configuration to only use setupFilesAfterEnv.
>
>  Configuration Documentation:
>  https://jestjs.io/docs/configuration.html