@youknowriad 2019-03-19
It's a good start. Other persons could chime-in and try to improve things.