@ellatrix 2019-03-19
`master`:

```
Average time to load: 4770.4ms
Average time to DOM content load: 3960.8ms
Average time to type character: 60.96ms
Slowest time to type character: 130ms
Fastest time to type character: 47ms
```

`5.2.0`:

Average time to load: 4742.7ms
Average time to DOM content load: 3928.4ms
Average time to type character: 63.07ms
Slowest time to type character: 132ms
Fastest time to type character: 48ms

`5.1.0`:

```
Average time to load: 5851.9ms
Average time to DOM content load: 4299.9ms
Average time to type character: 42.21ms
Slowest time to type character: 127ms
Fastest time to type character: 30ms
```

`5.0.0`:

```
Average time to load: 5647.9ms
Average time to DOM content load: 4110.5ms
Average time to type character: 46.09ms
Slowest time to type character: 87ms
Fastest time to type character: 38ms
```

`4.9.0`:

```
Average time to load: 7290.7ms
Average time to DOM content load: 5250.8ms
Average time to type character: 49.08ms
Slowest time to type character: 88ms
Fastest time to type character: 37ms
```