@ellatrix 2019-03-19
Had to tweak some files to test the other versions:

`4.8.0`:

```
Average time to load: 7445.2ms
Average time to DOM content load: 5752.4ms
Average time to type character: 129.305ms
Slowest time to type character: 264ms
Fastest time to type character: 110ms
```

`4.7.0`:

```
Average time to load: 8772.5ms
Average time to DOM content load: 6898.1ms
Average time to type character: 136.305ms
Slowest time to type character: 287ms
Fastest time to type character: 120ms
```