@aduth 2019-03-19
Re: https://github.com/WordPress/gutenberg/pull/14506#issuecomment-474324498

I'd imagine the warning may be a result of the recent Jest version upgrade, since your script does not account for the need to run `npm install` after checking out the tag.