@ellatrix 2019-03-19
@aduth Thanks :) I figured it out while manually checking out the tags and running the tests. Still, the whole idea is kind of flawed, as it only works back to 4.9.0. Earlier there wasn't a separate e2e package. It would be better to do something like @youknowriad suggested, but it doesn't seem like a small task. The current PR is still useable though, we can run it for every version we release. I ran it back until 4.7.0 as you can see above. the results look good: 4.7.0 was very slow, we made some small improvements for 4.8.0, then 4.9.0 is significantly faster thanks to @youknowriad's PR, performance decreased again with 5.2.0, but `master` is still similar (no big increase in the coming 5.3.0 release).