@nerrad 2019-03-19
I wonder if there'd be some value in having a job dedicated to performance regression in travis?  So for instance if the variance is +100ms or something then the job would fail.  That might help catch pulls that significantly impact performance?  I'm not sure offhand how this would work, but presumably there would need to be some service that marks/saves the current performance metrics on the master branch as the baseline for pulls (then maybe those metrics could be a simple json object saved to a file in the repo? - or amazon s3)?