@ellatrix 2019-03-19
@nerrad Yeah, that would be cool. It would require the performance test to run all the time as well, probably best as a separate job. We can keep a file with the times to test against, hopefully lowering them as performance improvements are made.