@gziolo 2019-03-20
I can help with the follow-up PR if we decide to integrate it with Travis. We would have to find a way to put this benchmark into its own Jest run (maybe with projects feature?).