@ellatrix 2019-04-03
These branches can be used for 4.7 and 4.8:

https://github.com/WordPress/gutenberg/compare/v4.7.0...test/performance-4-7
https://github.com/WordPress/gutenberg/compare/v4.8.0...test/performance-4-8

They require the files to be moved. You can't just stash and checkout a different release.

Would be good to create a solution that can just install any Gutenberg version and run the tests, but at the moment I like to spend my time on other things (I don't think it's a trivial thing to do). Anyone is free to try of course.