@youknowriad 2019-04-16
I pushed a commit to be able to run the performance test using a simple CLI script

```
npm run test-performance
```

I think we should merge this branch.

My last question is should we keep the "neuralink" post or replace it with a similar lipsum post in terms of content/blocks (thinking about copyright)