@aduth 2019-05-01
> They require the files to be moved. You can't just stash and checkout a different release.

If I understand the issue here to be the need to stash new files, you can pass the `-u` flag.

```
git stash -u
```

https://stackoverflow.com/questions/835501/how-do-you-stash-an-untracked-file