@youknowriad 2019-05-24
I think we should merge this branch. I keep looking for it to run performance tests.