@aduth 2019-05-24
I've pushed an alternate test fixture document, ideally roughly equivalent in stress testing a variety of different contents (headers, lists, quotes, formatted paragraphs, images).

&nbsp;|Words|Headings|Paragraphs|Blocks
--|--|--|--|--
Before|36396|0|840|1065
After|38819|202|547|952