@youknowriad 2019-05-28
I noticed that with the recent changes, running the performance tests is a lot slower even if the results are similar. I wonder if it's because of the time needed to load the images. I don't think this should stop us from landing the PR but it would be good to mock these or load local images.