@aduth 2019-05-28
We could consider using `data:` URIs, like what we do with E2E block fixtures:

https://github.com/WordPress/gutenberg/blob/c0ac81a6465e15125f1b54ee2b1bf63076d21e62/packages/e2e-tests/fixtures/blocks/core__image.html